+++ crates/matrix-sdk/src/room/timeline/event_handler.rs
@@ -43,9 +43,9 @@ use tracing::{debug, error, field::debug, info, instrument, trace, warn};
 
 use super::{
     event_item::{
-        AnyOtherFullStateEventContent, BundledReactions, EventSendState, LocalEventTimelineItem,
-        MemberProfileChange, OtherState, Profile, RemoteEventTimelineItem, RoomMembershipChange,
-        Sticker,
+        AnyOtherFullStateEventContent, BundledReactions, EventSendState, EventTimelineItemKind,
+        LocalEventTimelineItem, MemberProfileChange, OtherState, Profile, RemoteEventOrigin,
+        RemoteEventTimelineItem, RoomMembershipChange, Sticker,
     },
     find_read_marker,
     read_receipts::maybe_add_implicit_read_receipt,
@@ -58,12 +58,10 @@ use crate::{events::SyncTimelineEventWithoutContent, room::timeline::MembershipC
 pub(super) enum Flow {
     Local {
         txn_id: OwnedTransactionId,
-        timestamp: MilliSecondsSinceUnixEpoch,
     },
     Remote {
         event_id: OwnedEventId,
         txn_id: Option<OwnedTransactionId>,
-        origin_server_ts: MilliSecondsSinceUnixEpoch,
         raw_event: Raw<AnySyncTimelineEvent>,
         position: TimelineItemPosition,
     },
@@ -72,6 +70,7 @@ pub(super) enum Flow {
 pub(super) struct TimelineEventMetadata {
     pub(super) sender: OwnedUserId,
     pub(super) sender_profile: Option<Profile>,
+    pub(super) timestamp: MilliSecondsSinceUnixEpoch,
     pub(super) is_own_event: bool,
     pub(super) encryption_info: Option<EncryptionInfo>,
     pub(super) read_receipts: IndexMap<OwnedUserId, Receipt>,
@@ -409,18 +408,18 @@ impl<'a> TimelineEventHandler<'a> {
         };
 
         if let Some((idx, event_item)) = rfind_event_by_id(self.items, event_id) {
-            let EventTimelineItem::Remote(remote_event_item) = event_item else {
+            let Some(remote_event_item) = event_item.as_remote() else {
                 error!("inconsistent state: reaction received on a non-remote event item");
                 return;
             };
 
             // Handling of reactions on redacted events is an open question.
             // For now, ignore reactions on redacted events like Element does.
-            if let TimelineItemContent::RedactedMessage = remote_event_item.content() {
+            if let TimelineItemContent::RedactedMessage = event_item.content() {
                 debug!("Ignoring reaction on redacted event");
                 return;
             } else {
-                let mut reactions = remote_event_item.reactions().clone();
+                let mut reactions = remote_event_item.reactions.clone();
                 let reaction_group = reactions.entry(c.relates_to.key.clone()).or_default();
 
                 if let Some(txn_id) = old_txn_id {
@@ -438,7 +437,7 @@ impl<'a> TimelineEventHandler<'a> {
                 self.items.set(
                     idx,
                     Arc::new(TimelineItem::Event(
-                        remote_event_item.with_reactions(reactions).into(),
+                        event_item.with_kind(remote_event_item.with_reactions(reactions)),
                     )),
                 );
                 self.result.items_updated += 1;
@@ -478,12 +477,12 @@ impl<'a> TimelineEventHandler<'a> {
     fn handle_redaction(&mut self, redacts: OwnedEventId, _content: RoomRedactionEventContent) {
         if let Some((_, rel)) = self.reaction_map.remove(&(None, Some(redacts.clone()))) {
             update_timeline_item!(self, &rel.event_id, "redaction", |event_item| {
-                let EventTimelineItem::Remote(remote_event_item) = event_item else {
+                let Some(remote_event_item) = event_item.as_remote() else {
                     error!("inconsistent state: reaction received on a non-remote event item");
                     return None;
                 };
 
-                let mut reactions = remote_event_item.reactions().clone();
+                let mut reactions = remote_event_item.reactions.clone();
 
                 let count = {
                     let Entry::Occupied(mut group_entry) = reactions.entry(rel.key.clone()) else {
@@ -507,7 +506,7 @@ impl<'a> TimelineEventHandler<'a> {
                 }
 
                 trace!("Removing reaction");
-                Some(remote_event_item.with_reactions(reactions).into())
+                Some(event_item.with_kind(remote_event_item.with_reactions(reactions)))
             });
 
             if !self.result.items_updated > 0 {
@@ -528,12 +527,12 @@ impl<'a> TimelineEventHandler<'a> {
         // `reaction_map`), it can still be present in the timeline items
         // directly with the raw event timeline feature (not yet implemented).
         update_timeline_item!(self, &redacts, "redaction", |event_item| {
-            let EventTimelineItem::Remote(remote_event_item) = event_item else {
-                    error!("inconsistent state: reaction received on a non-remote event item");
-                    return None;
-                };
+            let Some(remote_event_item) = event_item.as_remote() else {
+                error!("inconsistent state: reaction received on a non-remote event item");
+                return None;
+            };
 
-            Some(remote_event_item.to_redacted().into())
+            Some(event_item.with_kind(remote_event_item.to_redacted()))
         });
 
         if self.result.items_updated > 0 {
@@ -549,20 +548,17 @@ impl<'a> TimelineEventHandler<'a> {
         let NewEventTimelineItem { content } = item;
         let sender = self.meta.sender.to_owned();
         let sender_profile = TimelineDetails::from_initial_value(self.meta.sender_profile.clone());
+        let timestamp = self.meta.timestamp;
         let mut reactions = self.pending_reactions().unwrap_or_default();
 
-        let mut item = match &self.flow {
-            Flow::Local { txn_id, timestamp } => {
-                EventTimelineItem::Local(LocalEventTimelineItem::new(
-                    EventSendState::NotSentYet,
-                    txn_id.to_owned(),
-                    sender,
-                    sender_profile,
-                    *timestamp,
-                    content,
-                ))
+        let kind: EventTimelineItemKind = match &self.flow {
+            Flow::Local { txn_id } => {
+                let send_state = EventSendState::NotSentYet;
+                let transaction_id = txn_id.to_owned();
+                LocalEventTimelineItem { send_state, transaction_id }
             }
-            Flow::Remote { event_id, origin_server_ts, raw_event, .. } => {
+            .into(),
+            Flow::Remote { event_id, raw_event, position, .. } => {
                 // Drop pending reactions if the message is redacted.
                 if let TimelineItemContent::RedactedMessage = content {
                     if !reactions.is_empty() {
@@ -570,24 +566,39 @@ impl<'a> TimelineEventHandler<'a> {
                     }
                 }
 
-                EventTimelineItem::Remote(RemoteEventTimelineItem::new(
-                    event_id.clone(),
-                    sender,
-                    sender_profile,
-                    *origin_server_ts,
-                    content,
+                let origin = match position {
+                    TimelineItemPosition::Start => RemoteEventOrigin::Pagination,
+                    // We only paginate backwards for now, so End only happens for syncs
+                    TimelineItemPosition::End => RemoteEventOrigin::Sync,
+                    #[cfg(feature = "e2e-encryption")]
+                    TimelineItemPosition::Update(idx) => self.items[*idx]
+                        .as_event()
+                        .and_then(|ev| Some(ev.as_remote()?.origin))
+                        .unwrap_or_else(|| {
+                            error!("Decryption retried on a local event");
+                            RemoteEventOrigin::Unknown
+                        }),
+                };
+
+                RemoteEventTimelineItem {
+                    event_id: event_id.clone(),
                     reactions,
-                    self.meta.read_receipts.clone(),
-                    self.meta.is_own_event,
-                    self.meta.encryption_info.clone(),
-                    raw_event.clone(),
-                    self.meta.is_highlighted,
-                ))
+                    read_receipts: self.meta.read_receipts.clone(),
+                    is_own: self.meta.is_own_event,
+                    is_highlighted: self.meta.is_highlighted,
+                    encryption_info: self.meta.encryption_info.clone(),
+                    original_json: raw_event.clone(),
+                    latest_edit_json: None,
+                    origin,
+                }
+                .into()
             }
         };
 
+        let mut item = EventTimelineItem::new(sender, sender_profile, timestamp, content, kind);
+
         match &self.flow {
-            Flow::Local { timestamp, .. } => {
+            Flow::Local { .. } => {
                 trace!("Adding new local timeline item");
 
                 // Check if the latest event has the same date as this event.
@@ -596,7 +607,7 @@ impl<'a> TimelineEventHandler<'a> {
                     let old_ts = latest_event.timestamp();
 
                     if let Some(day_divider_item) =
-                        maybe_create_day_divider_from_timestamps(old_ts, *timestamp)
+                        maybe_create_day_divider_from_timestamps(old_ts, timestamp)
                     {
                         trace!("Adding day divider");
                         self.items.push_back(Arc::new(day_divider_item));
@@ -604,18 +615,13 @@ impl<'a> TimelineEventHandler<'a> {
                 } else {
                     // If there is no event item, there is no day divider yet.
                     trace!("Adding first day divider");
-                    self.items.push_back(Arc::new(TimelineItem::day_divider(*timestamp)));
+                    self.items.push_back(Arc::new(TimelineItem::day_divider(timestamp)));
                 }
 
                 self.items.push_back(Arc::new(item.into()));
             }
 
-            Flow::Remote {
-                position: TimelineItemPosition::Start,
-                event_id,
-                origin_server_ts,
-                ..
-            } => {
+            Flow::Remote { position: TimelineItemPosition::Start, event_id, .. } => {
                 if self
                     .items
                     .iter()
@@ -642,14 +648,13 @@ impl<'a> TimelineEventHandler<'a> {
                     self.items.get(offset).and_then(|item| item.as_virtual())
                 {
                     if let Some(day_divider_item) =
-                        maybe_create_day_divider_from_timestamps(*divider_ts, *origin_server_ts)
+                        maybe_create_day_divider_from_timestamps(*divider_ts, timestamp)
                     {
                         self.items.insert(offset, Arc::new(day_divider_item));
                     }
                 } else {
                     // The list must always start with a day divider.
-                    self.items
-                        .insert(offset, Arc::new(TimelineItem::day_divider(*origin_server_ts)));
+                    self.items.insert(offset, Arc::new(TimelineItem::day_divider(timestamp)));
                 }
 
                 if self.track_read_receipts {
@@ -665,20 +670,14 @@ impl<'a> TimelineEventHandler<'a> {
                 self.items.insert(offset + 1, Arc::new(item.into()));
             }
 
-            Flow::Remote {
-                position: TimelineItemPosition::End,
-                txn_id,
-                event_id,
-                origin_server_ts,
-                ..
-            } => {
+            Flow::Remote { position: TimelineItemPosition::End, txn_id, event_id, .. } => {
                 let result = rfind_event_item(self.items, |it| {
                     txn_id.is_some() && it.transaction_id() == txn_id.as_deref()
                         || it.event_id() == Some(event_id)
                 });
 
                 if let Some((idx, old_item)) = result {
-                    if let EventTimelineItem::Remote(old_item) = old_item {
+                    if let Some(old_item) = old_item.as_remote() {
                         // Item was previously received from the server. This
                         // should be very rare normally, but with the sliding-
                         // sync proxy, it is actually very common.
@@ -698,8 +697,7 @@ impl<'a> TimelineEventHandler<'a> {
                     //       old and new item?
 
                     if idx == self.items.len() - 1
-                        && timestamp_to_date(old_item.timestamp())
-                            == timestamp_to_date(*origin_server_ts)
+                        && timestamp_to_date(old_item.timestamp()) == timestamp_to_date(timestamp)
                     {
                         // If the old item is the last one and no day divider
                         // changes need to happen, replace and return early.
@@ -760,7 +758,7 @@ impl<'a> TimelineEventHandler<'a> {
                     let old_ts = latest_event.timestamp();
 
                     if let Some(day_divider_item) =
-                        maybe_create_day_divider_from_timestamps(old_ts, *origin_server_ts)
+                        maybe_create_day_divider_from_timestamps(old_ts, timestamp)
                     {
                         trace!("Adding day divider");
                         self.items.push_back(Arc::new(day_divider_item));
@@ -768,7 +766,7 @@ impl<'a> TimelineEventHandler<'a> {
                 } else {
                     // If there is no event item, there is no day divider yet.
                     trace!("Adding first day divider");
-                    self.items.push_back(Arc::new(TimelineItem::day_divider(*origin_server_ts)));
+                    self.items.push_back(Arc::new(TimelineItem::day_divider(timestamp)));
                 }
 
                 if self.track_read_receipts {