+++ crates/matrix-sdk/src/room/timeline/tests/invalid.rs
@@ -3,7 +3,7 @@ use eyeball_im::VectorDiff;
 use futures_util::StreamExt;
 use matrix_sdk_test::async_test;
 use ruma::{
-    assign, event_id,
+    assign,
     events::{
         relation::Replacement,
         room::message::{self, MessageType, RoomMessageEventContent},
@@ -25,11 +25,11 @@ async fn invalid_edit() {
     let _day_divider =
         assert_matches!(stream.next().await, Some(VectorDiff::PushBack { value }) => value);
     let item = assert_matches!(stream.next().await, Some(VectorDiff::PushBack { value }) => value);
-    let event = item.as_event().unwrap().as_remote().unwrap();
+    let event = item.as_event().unwrap();
     let msg = event.content().as_message().unwrap();
     assert_eq!(msg.body(), "test");
 
-    let msg_event_id = event.event_id();
+    let msg_event_id = event.event_id().unwrap();
 
     let edit = assign!(RoomMessageEventContent::text_plain(" * fake"), {
         relates_to: Some(message::Relation::Replacement(Replacement::new(
@@ -65,9 +65,9 @@ async fn invalid_event_content() {
     let _day_divider =
         assert_matches!(stream.next().await, Some(VectorDiff::PushBack { value }) => value);
     let item = assert_matches!(stream.next().await, Some(VectorDiff::PushBack { value }) => value);
-    let event_item = item.as_event().unwrap().as_remote().unwrap();
+    let event_item = item.as_event().unwrap();
     assert_eq!(event_item.sender(), "@alice:example.org");
-    assert_eq!(event_item.event_id(), event_id!("$eeG0HA0FAZ37wP8kXlNkxx3I").to_owned());
+    assert_eq!(event_item.event_id().unwrap(), "$eeG0HA0FAZ37wP8kXlNkxx3I");
     assert_eq!(event_item.timestamp(), MilliSecondsSinceUnixEpoch(uint!(10)));
     let event_type = assert_matches!(
         event_item.content(),
@@ -89,9 +89,9 @@ async fn invalid_event_content() {
         .await;
 
     let item = assert_matches!(stream.next().await, Some(VectorDiff::PushBack { value }) => value);
-    let event_item = item.as_event().unwrap().as_remote().unwrap();
+    let event_item = item.as_event().unwrap();
     assert_eq!(event_item.sender(), "@alice:example.org");
-    assert_eq!(event_item.event_id(), event_id!("$d5G0HA0FAZ37wP8kXlNkxx3I").to_owned());
+    assert_eq!(event_item.event_id().unwrap(), "$d5G0HA0FAZ37wP8kXlNkxx3I");
     assert_eq!(event_item.timestamp(), MilliSecondsSinceUnixEpoch(uint!(2179)));
     let (event_type, state_key) = assert_matches!(
         event_item.content(),