+++ crates/matrix-sdk/src/sliding_sync/room.rs
@@ -25,7 +25,7 @@ use crate::{
     Client,
 };
 
-/// Room details, provided by a [`SlidingSync`] instance.
+/// Room details, provided by a [`SlidingSync`][super::SlidingSync] instance.
 #[derive(Debug, Clone)]
 pub struct SlidingSyncRoom {
     client: Client,
@@ -100,7 +100,7 @@ impl SlidingSyncRoom {
     ///
     /// Use `Timeline::latest_event` instead if you already have a timeline for
     /// this `SlidingSyncRoom`.
-    #[instrument(skip_all, parent = &self.client.inner.root_span)]
+    #[instrument(skip_all)]
     pub async fn latest_event(&self) -> Option<EventTimelineItem> {
         self.timeline_builder()?.build().await.latest_event().await
     }