+++ testing/sliding-sync-integration-test/src/lib.rs
@@ -10,7 +10,6 @@ use assert_matches::assert_matches;
 use eyeball_im::VectorDiff;
 use futures::{pin_mut, stream::StreamExt};
 use matrix_sdk::{
-    room::timeline::EventTimelineItem,
     ruma::{
         api::client::{
             error::ErrorKind as RumaError,
@@ -224,17 +223,17 @@ async fn modifying_timeline_limit() -> anyhow::Result<()> {
         assert_matches!(timeline_items[0].as_virtual(), Some(_));
 
         // Second timeline item.
-        let latest_remote_event = assert_matches!(
-            timeline_items[1].as_event(),
-            Some(EventTimelineItem::Remote(remote_event)) => remote_event
-        );
-        all_event_ids.push(latest_remote_event.event_id().to_owned());
+        let latest_remote_event = timeline_items[1].as_event().unwrap();
+        all_event_ids.push(latest_remote_event.event_id().unwrap().to_owned());
 
         // Test the room to see the last event.
-        assert_matches!(room.latest_event().await, Some(EventTimelineItem::Remote(remote_event)) => {
-            assert_eq!(remote_event.event_id(), latest_remote_event.event_id(), "Unexpected latest event");
-            assert_eq!(remote_event.content().as_message().unwrap().body(), "Message #19");
-        });
+        let latest_event = room.latest_event().await.unwrap();
+        assert_eq!(
+            latest_event.event_id(),
+            latest_remote_event.event_id(),
+            "Unexpected latest event"
+        );
+        assert_eq!(latest_event.content().as_message().unwrap().body(), "Message #19");
 
         (room, timeline, timeline_stream)
     };
@@ -270,20 +269,19 @@ async fn modifying_timeline_limit() -> anyhow::Result<()> {
 
         // The first 19th items are `VectorDiff::PushBack`.
         for nth in 0..19 {
-            assert_matches!(timeline_stream.next().await, Some(VectorDiff::PushBack { value }) => {
-                let remote_event = assert_matches!(
-                    value.as_event(),
-                    Some(EventTimelineItem::Remote(remote_event)) => remote_event
-                );
+            let value = assert_matches!(
+                timeline_stream.next().await,
+                Some(VectorDiff::PushBack { value }) => value
+            );
+            let event = value.as_event().unwrap();
 
-                // Check messages arrived in the correct order.
-                assert_eq!(
-                    remote_event.content().as_message().expect("Received event is not a message").body(),
-                    format!("Message #{nth}"),
-                );
+            // Check messages arrived in the correct order.
+            assert_eq!(
+                event.content().as_message().expect("Received event is not a message").body(),
+                format!("Message #{nth}"),
+            );
 
-                all_event_ids.push(remote_event.event_id().to_owned());
-            });
+            all_event_ids.push(event.event_id().unwrap().to_owned());
         }
 
         // The 20th item is a `VectorDiff::Remove`, i.e. the first message is removed.
@@ -294,22 +292,21 @@ async fn modifying_timeline_limit() -> anyhow::Result<()> {
 
         // And now, the initial message is pushed at the bottom, so the 21th item is a
         // `VectorDiff::PushBack`.
-        let latest_remote_event = assert_matches!(timeline_stream.next().await, Some(VectorDiff::PushBack { value }) => {
-            let remote_event = assert_matches!(
-                value.as_event(),
-                Some(EventTimelineItem::Remote(remote_event)) => remote_event
-            );
-            assert_eq!(remote_event.content().as_message().unwrap().body(), "Message #19");
-            assert_eq!(remote_event.event_id(), all_event_ids[0]);
+        let latest_remote_event = assert_matches!(
+            timeline_stream.next().await,
+            Some(VectorDiff::PushBack { value }) => value
+        );
 
-            remote_event.clone()
-        });
+        let event = latest_remote_event.as_event().unwrap();
+        let event_id = event.event_id().unwrap();
+        assert_eq!(event.content().as_message().unwrap().body(), "Message #19");
+        assert_eq!(event_id, all_event_ids[0]);
 
         // Test the room to see the last event.
-        assert_matches!(room.latest_event().await, Some(EventTimelineItem::Remote(remote_event)) => {
-            assert_eq!(remote_event.content().as_message().unwrap().body(), "Message #19");
-            assert_eq!(remote_event.event_id(), latest_remote_event.event_id(), "Unexpected latest event");
-        });
+        let latest_event = room.latest_event().await.unwrap();
+
+        assert_eq!(event.content().as_message().unwrap().body(), "Message #19");
+        assert_eq!(latest_event.event_id().unwrap(), event_id, "Unexpected latest event");
 
         // Ensure there is no event ID duplication.
         {
@@ -865,7 +862,7 @@ async fn fast_unfreeze() -> anyhow::Result<()> {
         let (sliding_window_list, growing_sync) = build_lists()?;
         let sync_proxy = sync_proxy_builder
             .clone()
-            .storage_key(Some("sliding_sync".to_string()))
+            .storage_key(Some("sliding_sync".to_owned()))
             .add_list(sliding_window_list)
             .add_list(growing_sync)
             .build()
@@ -891,7 +888,7 @@ async fn fast_unfreeze() -> anyhow::Result<()> {
     let start = Instant::now();
     let _sync_proxy = sync_proxy_builder
         .clone()
-        .storage_key(Some("sliding_sync".to_string()))
+        .storage_key(Some("sliding_sync".to_owned()))
         .add_list(sliding_window_list)
         .add_list(growing_sync)
         .build()
@@ -962,7 +959,7 @@ async fn continue_on_reset() -> anyhow::Result<()> {
     println!("starting the sliding sync setup");
     let sync_proxy = sync_proxy_builder
         .clone()
-        .storage_key(Some("sliding_sync".to_string()))
+        .storage_key(Some("sliding_sync".to_owned()))
         .add_list(growing_sync)
         .build()
         .await?;
@@ -1047,7 +1044,7 @@ async fn noticing_new_rooms_in_growing() -> anyhow::Result<()> {
     println!("starting the sliding sync setup");
     let sync_proxy = sync_proxy_builder
         .clone()
-        .storage_key(Some("sliding_sync".to_string()))
+        .storage_key(Some("sliding_sync".to_owned()))
         .add_list(growing_sync)
         .build()
         .await?;