+++ bindings/matrix-sdk-crypto-js/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.1.0-alpha.7
+
+-   Add new accessors `Device.algorithms` and `Device.isSignedByOwner`
+-   In `OlmMachine.getUserDevices`, wait a limited time for any in-flight
+    device-list updates to complete.
+
 # v0.1.0-alpha.6
 
 -   Add new accessor `InboundGroupSession.senderKey`.